custom action to the toolbar without firing the editor blur event.
i wanted to add other custom dialog actions to the toolbar, the problem
is that :
data-wysihtml5-dialog-action="cancel"
data-wysihtml5-dialog-action="save" etc … unobserve the blur event to
(wisely) not fire the blur event when the mouse click on those
'buttons'/actions

if you want to add some custom dialog actions, the "unobserve blur" is
not fired. So it fires a blur event form the text editor and (in my
case) close the toolbar dialog.
so i added a custom:command event for custom dialog action like this:

data-wysihtml5-command=""

You can now add custom action to the toolbar without firing the editor
blur event.